Added UI themes and dedicated menus
Added dedicated menus to both the Host Session button and the Join Session button. The Host Session button displays a list of text boxes asking for the player's username and the port they want to receive data from. The Join session button displays a list of text boxes asking for the player's username, the opponent's IP address and port. Also applied a nicer theme to the UI.